README change, modularized function from Signup component, added axios and added communication with backend in Signup component
