+++ README.md
@@ -1,6 +1,11 @@
-# music-app-front
+# Music Application (frontend)
 Front end for the music app
 
+## Intent
+The purpose of this application is to create an application where users can listen to any song they wish to, via a search bar. We will also add the ability for users to login with authentication, and likewise, the ability for each individual user to store their own playlists they create themselves.
 
 ## link for outline
 https://keian.invisionapp.com/freehand/Music-Application-8NxQ7ZAWV
+
+## Backend link
+[View the backend here!](https://github.com/Keian-A/music-app-back)
